Update README to focus on ad timer capabilities
This PR updates the README.md to focus specifically on the ad timer fast-forwarding capabilities of the extension, removing mentions of other uses like videos, animations, or games. The README now emphasizes how the extension helps users bypass waiting timers on advertising websites, which was the primary purpose of creating this extension.